ref: moved module configurations and function definition into its own service
